requirements.txt
discord-py-interactions
discord-typings